addProduct

+++ app/Http/Controllers/Admin/HomeController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers\Admin;
 
 use App\Http\Controllers\Controller;
 use App\Http\Requests\Admin\PostRequest;
+use App\Models\Users;
 use App\Repositories\User\UserReponsitoryInterface;
 use Illuminate\Http\Request;
 
@@ -56,7 +57,7 @@ class HomeController extends Controller
             'password' => bcrypt('1'),
             'reset_password' => bcrypt('1'),
             'status' => "active",
-            'avatar' =>$pathAvatar,
+            'avatar' => $pathAvatar,
             'flag_delete' => 0,
             'created_at' => date(now('Asia/Ho_Chi_Minh')),
         ];
@@ -75,13 +76,16 @@ class HomeController extends Controller
     public function getEdit(Request $request)
     {
         $id = $request->get('id');
-        $getUser = $this->userRepository->getUser($id);
-        $data['getUser'] = $getUser;
-        if (!empty($getUser)) {
-            $request->session()->put('id', $id);
-            return view('admin.edit', $data);
+        if (Users::where('id', $id)->exists()) {
+            $getUser = $this->userRepository->getUser($id);
+            $data['getUser'] = $getUser;
+            if (!empty($getUser)) {
+                $request->session()->put('id', $id);
+                return view('admin.edit', $data);
+            }
+        } else {
+            return redirect()->route('admin.home');
         }
-        return redirect()->route('admin.home');
     }
 
     /** handle EditUser
@@ -103,7 +107,7 @@ class HomeController extends Controller
             'birthday' => $request->birthday,
             'first_name' => $request->first_name,
             'last_name' => $request->last_name,
-            'avatar' =>$pathAvatar,
+            'avatar' => $pathAvatar,
             'updated_at' => date(now('Asia/Ho_Chi_Minh')),
         ];
         $status = $this->userRepository->updateUser($dataUpdate, $id);
@@ -121,9 +125,14 @@ class HomeController extends Controller
     public function deleteUser(Request $request)
     {
         $id = $request->get('id');
-        $status = $this->userRepository->deleteUser($id);
-        if ($status) {
-            return redirect()->route("admin.home")->with("successDelete", __('messages.success.deleteUser'));
+        if (Users::where('id', $id)->exists()) {
+            $status = $this->userRepository->deleteUser($id);
+            if ($status) {
+                return redirect()->route("admin.home")->with("successDelete", __('messages.success.deleteUser'));
+            }
+        } else {
+            return redirect()->route('admin.home');
         }
+
     }
 }
